refactor project

+++ mysql/functions.php
@@ -1,5 +1,25 @@
 <?php include "db.php";?>
 <?php
+function createRows(){
+    if (isset($_POST['submit'])) {
+        global $connection;
+        $username = $_POST['username'];
+        $password = $_POST['password'];
+
+        $query = "INSERT INTO users(username,password)";
+        $query .= "VALUES ('$username','$password')";
+
+        $result = mysqli_query($connection, $query);
+        if (!$result) {
+            die('Query FAILED' . mysqli_error());
+        }else{
+            echo "Record Create";
+        }
+    }
+
+}
+
+
 
 function showAllData()
 {
@@ -23,39 +43,46 @@ function showAllData()
 
 function updateTable()
 {
-    global $connection;
-    $username = $_POST['username'];
-    $password = $_POST['password'];
-    $id = $_POST['id'];
+    if (isset($_POST['submit'])) {
+        global $connection;
+        $username = $_POST['username'];
+        $password = $_POST['password'];
+        $id = $_POST['id'];
 
-    $query = "UPDATE users SET ";
-    $query .= "username = '$username', ";
-    $query .= "password = '$password' ";
-    $query .= "WHERE id = $id ";
+        $query = "UPDATE users SET ";
+        $query .= "username = '$username', ";
+        $query .= "password = '$password' ";
+        $query .= "WHERE id = $id ";
 
-    $result = mysqli_query($connection, $query);
-    if (!$result) {
-        die("QUERY FAILED" . mysqli_error($connection));
+        $result = mysqli_query($connection, $query);
+        if (!$result) {
+            die("QUERY FAILED" . mysqli_error($connection));
+        } else {
+
+            echo "Record Updated";
+
+        }
     }
-
 }
-
 
 function deleteRows()
 {
-    global $connection;
-    $username = $_POST['username'];
-    $password = $_POST['password'];
-    $id = $_POST['id'];
+    if (isset($_POST['submit'])) {
+        global $connection;
+        $username = $_POST['username'];
+        $password = $_POST['password'];
+        $id = $_POST['id'];
 
-    $query = "DELETE FROM users ";
-    $query .= "WHERE id = $id ";
+        $query = "DELETE FROM users ";
+        $query .= "WHERE id = $id ";
 
-    $result = mysqli_query($connection, $query);
-    if (!$result) {
-        die("QUERY FAILED" . mysqli_error($connection));
+        $result = mysqli_query($connection, $query);
+        if (!$result) {
+            die("QUERY FAILED" . mysqli_error($connection));
+        } else {
+            echo "Record Deleted";
+        }
+
     }
-
 }
-
